Add wgpu_canvas_import_webgpu_context and wgpu_import_device
These functions allow a WebGPU context and device that were acquired in JS to be imported and used on the C++ side. This is useful in an application that might have a heavy amount of WebAssembly and wants to acquire a context and get something rendered to it before the wasm artifact has had time to load.