Promises have a behavior that if the .then() handler throws an exception, then the .catch() handler is invoked.
The intended design of the WebGPU .then/.catch callbacks is to invoke the callback only once, depending on success or failure. So we must guard all JS exception from the success handler so that it won't propagate to invoking the failure handler.

To do that, wrap the .then() handlers with a function that stops propagation of any JS exceptions that the then handlers might throw. (These exceptions could be the Emscripten 'unwind' or 'ExitStatus' handlers, that are thrown just to unwind, and not to signify any failure)